+++ src/main/kotlin/com/displee/cache/index/Index255.kt
@@ -1,0 +1,13 @@
+package com.displee.cache.index
+
+import com.displee.cache.CacheLibrary
+import com.displee.cache.ProgressListener
+import java.io.RandomAccessFile
+
+class Index255(origin: CacheLibrary, raf: RandomAccessFile) : Index(origin, 255, raf) {
+
+    override fun update(xteas: Map<Int, IntArray>, listener: ProgressListener?): Boolean {
+        throw IllegalAccessException("Not allowed.")
+    }
+
+}
